backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.0.0",
    "numpy>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.27.2",
]
